requirements.txt
attrs==16.3.0
click==6.7
PTable==0.9.2
pyaml==16.12.2
PyYAML==3.12
flask-sqlalchemy==2.2
flask-admin==1.5.0 
wtforms==2.1